server/requirements.txt
fastapi
jinja2
pydantic
pyyaml
uvicorn
